Update project files

+++ scripts/Reset-Examples.ps1
@@ -1,6 +1,7 @@
 git checkout origin/main examples
-# TODO: if ./examples/faceless then remove
-Remove-Item -Path ./examples/faceless -Recurse -Force
+if (Test-Path -Path ./examples/faceless) {
+  Remove-Item -Path ./examples/faceless -Recurse -Force
+}
 if (Test-Path -Path ./examples/labels) {
   Remove-Item -Path ./examples/labels -Recurse -Force
 }
